# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Eduator Clean Backend API
  version: 0.4.0
  description: >-
    Clean backend API contract for local PostgreSQL + JWT + HTTP API keys (baseline **0.4.0**).
    Includes OpenRouter multi-model AI (migration `018_ai_provider_openrouter.sql`),
    pgvector-backed RAG (migration `015_pgvector_document_chunks.sql`), user HTTP API keys,
    usage analytics (`api_access_log` after migration `006_api_access_log.sql`),
    and first-party traffic exclusion via `X-Eduator-Client` (see docs).
servers:
  - url: http://localhost:4000
    description: Local backend
...
paths:
...
  /v1/users:
    get:
      tags: [Users]
      summary: List users
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User list
  /v1/users/{id}/password:
    patch:
      tags: [Users]
      summary: Update user password (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password updated
        '403':
          description: Forbidden
        '404':
          description: User not found
...
  /v1/admin/users:
    post:
      tags: [Admin]
      summary: Create an operator or user account
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: User created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...